move subProcess to baseCommand
