+++ packages/@tinacms/cli/src/next/commands/baseCommands.ts
@@ -1,4 +1,9 @@
 import { Command, Option } from 'clipanion'
+import chalk from 'chalk'
+
+import type { ChildProcess } from 'child_process'
+import { startSubprocess2 } from '../../utils/start-subprocess'
+import { logger } from '../../logger'
 
 /**
  * Base Command for Dev and build
@@ -35,5 +40,45 @@ export abstract class BaseCommand extends Command {
   noTelemetry = Option.Boolean('--noTelemetry', false, {
     description: 'Disable anonymous telemetry that is collected',
   })
+
   abstract execute(): Promise<number | void>
+
+  async startSubCommand() {
+    let subProc: ChildProcess | undefined
+    if (this.subCommand) {
+      subProc = await startSubprocess2({ command: this.subCommand })
+      logger.info(`Starting subprocess: ${chalk.cyan(this.subCommand)}`)
+    }
+    function exitHandler(options, exitCode) {
+      if (subProc) {
+        subProc.kill()
+      }
+      process.exit()
+    }
+    //do something when app is closing
+    process.on('exit', exitHandler)
+    //catches ctrl+c event
+    process.on('SIGINT', exitHandler)
+    // catches "kill pid" (for example: nodemon restart)
+    process.on('SIGUSR1', exitHandler)
+    process.on('SIGUSR2', exitHandler)
+    //catches uncaught exceptions
+    process.on('uncaughtException', exitHandler)
+  }
+
+  logDeprecationWarnings() {
+    if (this.isomorphicGitBridge) {
+      logger.warn('--isomorphicGitBridge has been deprecated')
+    }
+    if (this.experimentalDataLayer) {
+      logger.warn(
+        '--experimentalDataLayer has been deprecated, the data layer is now built-in automatically'
+      )
+    }
+    if (this.noSDK) {
+      logger.warn(
+        '--noSDK has been deprecated, and will be unsupported in a future release. This should be set in the config at client.skip = true'
+      )
+    }
+  }
 }
